+++ src/main/java/xyz/veiasai/mongodb/dao/OrderRepository.java
@@ -5,9 +5,10 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import xyz.veiasai.mongodb.pojo.Commodity;
 import xyz.veiasai.mongodb.pojo.Order;
 
+import java.math.BigInteger;
 import java.util.List;
 
-public interface OrderRepository extends MongoRepository<Order, Integer>{
+public interface OrderRepository extends MongoRepository<Order, BigInteger>{
 
     public Order findByIndex(Integer index);
